Deprecate SOAP
Deprecate the query() method and all methods that use it.

Start ignoring XML::Parser::Lite, which has been split off from
SOAP::Lite but which depends on it, and fails to install apparently due
to missing dependencies, and which does not seem to be much maintained,
since the most recent release was several months ago, and about half the
smoke tests fail.