Add support for https:
The choice of http: or https: is user-settable by the 'scheme'
attribute. This defaults to the value of the ASTRO_SIMBAD_CLIENT_SCHEME
environment variable, or to 'http' if that is not set or is invalid.

Thanks to Christian Carey for making me aware of this, and providing
patches. I would have used more of his actual code, but I took the
opportunity to do some refactoring of the URL logic.